Service Page #49
Added a service page where this website could offer different types of services such as: Group Support,
Counselling, 24/7 Customer Care. This Page is  seperate so as to add to the functionality and moto of the page.

Added features like contact on whatsapp, chatbot, through email, seek support groups and much more!